requirements.txt
acp-sdk>=0.8.1
PyGithub>=2.1.1
python-dotenv>=1.0.0
langchain>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.0.0
streamlit>=1.32.0 